Revert "Make role_id BigInt"
This reverts commit 27548bac24b20a19c31dfbe85c1420b845929546.

+++ discord_bots/models.py
@@ -1288,7 +1288,7 @@ class QueueRole:
         },
     )
     role_id: int = field(
-        metadata={"sa": Column(BigInteger, nullable=False)},
+        metadata={"sa": Column(Integer, nullable=False)},
     )
     id: str = field(
         init=False,
